Adjust test to avoid going outside viewport
This caused a problem when testing in Firefox when repeated moves
dragged the control point outside the viewport. It might be a problem to
look into some time, but I can't reproduce the issue and it wasn't
really the goal of this test to verify the behaviour when dragging
points outside the viewport.